# app/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/suggestions": {
      "get": {
        "operationId": "SuggestionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список предложений"
          }
        },
        "summary": "Список предложений с фильтрами и пагинацией",
        "tags": [
          "Suggestions"
        ]
      },
      "post": {
        "operationId": "SuggestionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSuggestionBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Созданное предложение"
          },
          "400": {
            "description": "X-User-Id required"
          }
        },
        "summary": "Создание предложения",
        "tags": [
          "Suggestions"
        ]
      }
    },
...
    "/api/v1/suggestions/{id}": {
      "get": {
        "operationId": "SuggestionsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Предложение"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Детальная панель предложения",
        "tags": [
          "Suggestions"
        ]
      },
      "patch": {
        "operationId": "SuggestionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSuggestionBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлённое предложение"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Обновление предложения (статус, категория)",
        "tags": [
          "Suggestions"
        ]
      },
      "delete": {
        "operationId": "SuggestionsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Удалено"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Удаление предложения",
        "tags": [
          "Suggestions"
        ]
      }
    },
    "/api/v1/suggestions/{id}/status": {
      "patch": {
        "operationId": "SuggestionsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSuggestionStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлённое предложение"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Изменение статуса предложения",
        "tags": [
          "Suggestions"
        ]
      }
    },
    "/api/v1/suggestions/merge": {
      "post": {
        "operationId": "SuggestionsController_merge",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Результат слияния"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Слияние предложений (sourceId → targetId)",
        "tags": [
          "Suggestions"
        ]
      }
    },
    "/api/v1/suggestions/{id}/vote": {
      "post": {
        "operationId": "SuggestionsController_vote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Результат голосования"
          },
          "400": {
            "description": "X-User-Id required"
          }
        },
        "summary": "Голосование upvote/downvote",
        "tags": [
          "Suggestions"
        ]
      }
    },
    "/api/v1/suggestions/{id}/discord-preview": {
      "get": {
        "operationId": "SuggestionsController_discordPreview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Embed для Discord"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Discord Preview: payload/embed для предпросмотра",
        "tags": [
          "Suggestions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Suggestions API",
    "description": "API для управления предложениями (suggestions)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateSuggestionBodyDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSuggestionBodyDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSuggestionStatusDto": {
        "type": "object",
        "properties": {}
      },
      "VoteDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}