# app/src/modules/suggestions/dto/update-suggestion-body.dto.ts
import { IsArray, IsOptional, IsString, IsUrl, MaxLength } from 'class-validator';

export class UpdateSuggestionBodyDto {
  @IsOptional()
  @IsString()
  @MaxLength(500)
  title?: string;

  @IsOptional()
  @IsString()
  @MaxLength(10000)
  description?: string;

  @IsOptional()
  @IsString()
  @MaxLength(100)
  category?: string;

  @IsOptional()
  @IsArray()
  @IsUrl({}, { each: true })
  images?: string[];
}